Restructure core read-in functions
